+++ x/aswap/msg.go
@@ -1,25 +1,18 @@
 package aswap
 
 import (
-	"github.com/iov-one/weave"
 	coin "github.com/iov-one/weave/coin"
 	"github.com/iov-one/weave/errors"
 	"github.com/iov-one/weave/migration"
 )
 
 func init() {
-	// Migration needs to be registered for every message introduced in the codec.
-	// This is the convention to message versioning.
 	migration.MustRegister(1, &CreateMsg{}, migration.NoModification)
 	migration.MustRegister(1, &ReleaseMsg{}, migration.NoModification)
 	migration.MustRegister(1, &ReturnSwapMsg{}, migration.NoModification)
 }
 
 const (
-	pathCreateSwap  = "aswap/create"
-	pathReleaseSwap = "aswap/release"
-	pathReturnSwap  = "aswap/return"
-
 	maxMemoSize int = 128
 	// preimage size in bytes
 	preimageSize int = 32
@@ -27,25 +20,9 @@ const (
 	preimageHashSize int = 32
 )
 
-var _ weave.Msg = (*CreateMsg)(nil)
-var _ weave.Msg = (*ReleaseMsg)(nil)
-var _ weave.Msg = (*ReturnSwapMsg)(nil)
-
-// ROUTING, Path method fulfills weave.Msg interface to allow routing
-
 func (CreateMsg) Path() string {
-	return pathCreateSwap
+	return "aswap/create"
 }
-
-func (ReleaseMsg) Path() string {
-	return pathReleaseSwap
-}
-
-func (ReturnSwapMsg) Path() string {
-	return pathReturnSwap
-}
-
-// VALIDATION, Validate method makes sure basic rules are enforced upon input data and fulfills weave.Msg interface
 
 func (m *CreateMsg) Validate() error {
 	if err := m.Metadata.Validate(); err != nil {
@@ -77,6 +54,10 @@ func (m *CreateMsg) Validate() error {
 	return err
 }
 
+func (ReleaseMsg) Path() string {
+	return "aswap/release"
+}
+
 func (m *ReleaseMsg) Validate() error {
 	if err := m.Metadata.Validate(); err != nil {
 		return errors.Wrap(err, "metadata")
@@ -90,6 +71,11 @@ func (m *ReleaseMsg) Validate() error {
 		return errors.Wrapf(errors.ErrInput, "preimage should be exactly %d byte long", preimageSize)
 	}
 	return nil
+}
+
+func (ReturnSwapMsg) Path() string {
+	return "aswap/return"
+
 }
 
 func (m *ReturnSwapMsg) Validate() error {