+++ CHANGELOG.md
@@ -3,11 +3,13 @@
 ## HEAD
 
 - `bnsd/x/username` genesis initializer implemented and included in `bnsd`.
-- Support gov proposal vote, deletion and tally in `bnscli` 
+- Support gov proposal vote, deletion and tally in `bnscli`
 
 Breaking changes
 
 - Unify all message paths to follow pattern `<package>/<message_name>`
+- `app.Router` interface was changed. Handler registration requires a message
+  and not message path.
 
 ## 0.17.0